Functionality for correspondence analysis ('ca') objects

These methods extract data from, and attribute new data to, objects of class "ca" from the ca package.

Usage

# S3 method for class 'ca'
as_tbl_ord(x)

# S3 method for class 'ca'
recover_rows(x)

# S3 method for class 'ca'
recover_cols(x)

# S3 method for class 'ca'
recover_inertia(x)

# S3 method for class 'ca'
recover_conference(x)

# S3 method for class 'ca'
recover_coord(x)

# S3 method for class 'ca'
recover_aug_rows(x)

# S3 method for class 'ca'
recover_aug_cols(x)

# S3 method for class 'ca'
recover_aug_coord(x)

Arguments

x

An ordination object.

Value

The recovery generics recover_*() return core model components, distribution of inertia, supplementary elements, and intrinsic metadata; but they require methods for each model class to tell them what these components are.

The generic as_tbl_ord() returns its input wrapped in the 'tbl_ord' class. Its methods determine what model classes it is allowed to wrap. It then provides 'tbl_ord' methods with access to the recoverers and hence to the model components.

Single, multiple, and joint CA

Greenacre (2007) proposes joint correspondence analysis (JCA) as a counterpart to classical multiple correspondence analysis (MCA). These, together with classical correspondence analysis (CA), are implemented in the ca package.

References

Greenacre MJ (2007) Correspondence Analysis in Practice, Second Edition. Boca Raton: Chapman & Hall/CRC, ISBN 978-1-58488-616-7. http://www.carme-n.org/?sec=books2

See also

Examples

# table of hair and eye color data collapsed by sex
class(HairEyeColor)
#> [1] "table"
haireye <- as.data.frame(rowSums(HairEyeColor, dims = 2L))
print(haireye)
#>       Brown Blue Hazel Green
#> Black    68   20    15     5
#> Brown   119   84    54    29
#> Red      26   17    14    14
#> Blond     7   94    10    16

if (require(ca)) {# {ca}

# use correspondence analysis to construct row and column profiles
haireye %>%
  ca::ca() %>%
  as_tbl_ord() %>%
  print() -> haireye_ca

# recover row and column profiles
get_rows(haireye_ca)
get_cols(haireye_ca)
# augment profiles with names, masses, distances, and inertias
(haireye_ca <- augment_ord(haireye_ca))

}# {ca}
#> Loading required package: ca
#> Warning: there is no package called ‘ca’